refactor: refactored text to own component
